ports: cast return result from getpid().
Fixes a warning caused by a type mismatch (int vs. pid_t).


git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@5881 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ libhb/ports.c
@@ -467,7 +467,7 @@ void hb_get_temporary_directory( char path[512] )
     if( base[strlen(base)-1] == '/' )
         base[strlen(base)-1] = '\0';
 
-    snprintf( path, 512, "%s/hb.%d", base, getpid() );
+    snprintf( path, 512, "%s/hb.%d", base, (int)getpid() );
 }
 
 /************************************************************************
